feature: Fixed bug: when adding to sys.path a bytestring should be added, not a unicode string.
This allows relative paths to be used for exec() later since the paths
are relative to entries in sys.path, and the traceback code will
look in sys.path for the file. Previously we got an error when linecache
was trying to join entries in sys.path with the filename, but now
both are bytestrings so it should all be good.

In Python 3 of course they will need to be unicode strings. Each encode/decode
line will need to become conditional on Python version.